# radio-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radiology API",
        "description": "Appointment booking application for a radiology department",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://radiologie-rdv.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Enregistrement d'un nouveau patient",
                "operationId": "573de1fed352c1205a32c4d1b9877375",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0612345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirection vers le dashboard après enregistrement"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/register/medecin": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Enregistrement d'un nouveau médecin",
                "operationId": "99df6c54e4f5d2d3903ce045d7f47fe7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. Marie Curie"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marie.curie@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0612345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirection vers la route de redirection par rôle"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Valider l'adresse email de l'utilisateur (vérification email)",
                "operationId": "defff15cf49d15083f3d7d2284955a29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Hash de vérification email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirection vers le dashboard avec paramètre 'verified=1' en cas de succès"
                    },
                    "403": {
                        "description": "Non autorisé, si l'email a déjà été vérifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}